Introduce refreshView method
When created for the first time we don't want to replace the view's content
until the new dom tree is ready.
However, using html multiple times resets the event handlers on the view's
el.

Thus we attache only once : When render is completed. To redraw the view
after a model update or a user event, use refreshView :
it will change the dom only below el and ensure event binding is preserved.